requirements.txt
tencentcloud_sdk_python_common==3.0.1186
tencentcloud_sdk_python_hunyuan==3.0.1186
# meme_generator==0.0.19.1
spark-ai-python==0.4.1
feedparser==6.0.11
wcferry==39.2.4.0
Requests==2.32.3
schedule==1.2.2
Pillow==10.4.0
urllib3==2.2.3
PyYAML==6.0.2
cprint==1.2.2
lxml==5.2.2